[Messenger] Fix float value for worker memory limit

+++ Command/ConsumeMessagesCommand.php
@@ -289,7 +289,7 @@ EOF
         } elseif (str_starts_with($max, '0')) {
             $max = \intval($max, 8);
         } else {
-            $max = (int) $max;
+            $max = (float) $max;
         }
 
         switch (substr(rtrim($memoryLimit, 'b'), -1)) {
@@ -302,6 +302,6 @@ EOF
             case 'k': $max *= 1024;
         }
 
-        return $max;
+        return (int) $max;
     }
 }
